Substantially update README.md
Updates the README.md to provide more background behind the SDK layout,
add a few git usage examples to help users new to submodules, and
improve some formatting.

Notably, IAR support was previously missing.  It continues to be poorly
documented, but at least a placeholder with a pointer to the Quick Start
Guides is now provided.

Signed-off-by: Chris Ring <c-ring@ti.com>